+++ containers/daap-python-base/src/var/task/versioning.py
@@ -79,22 +79,27 @@ class InvalidUpdate(Exception):
     """
 
 
-class VersionCreator:
+class VersionManager:
     """
     Service to create new versions of a data product when metadata or schema are updated.
     """
 
     def __init__(self, data_product_name, logger: DataPlatformLogger):
         self.data_product_config = DataProductConfig(name=data_product_name)
+        self.latest_version = self.data_product_config.latest_version
         self.logger = logger
 
     def update_metadata_remove_schemas(self, schema_list: list[str]) -> str:
         """Handles removing schema(s) for a data product."""
         s3_client = boto3.client("s3")
+        # remove any list duplicates and preserve list order
+        schema_list = list({k: None for k in schema_list}.keys())
+
+        data_product_name = self.data_product_config.name
 
         current_metadata = (
             DataProductMetadata(
-                data_product_name=self.data_product_config.name,
+                data_product_name=data_product_name,
                 logger=self.logger,
                 input_data=None,
             )
@@ -108,30 +113,31 @@ class VersionCreator:
             schema in current_schemas for schema in schema_list
         )
         if not valid_schemas_to_delete:
-            error = f"Invalid schemas found in schema_list: {schema_list}"
+            schemas_not_in_current = list(set(schema_list).difference(current_schemas))
+            error = f"Invalid schemas found in schema_list: {sorted(schemas_not_in_current)}"
             self.logger.error(error)
             raise InvalidUpdate(error)
 
         self.logger.info(f"schemas to delete: {schema_list}")
-        for schema in schema_list:
+        for schema_name in schema_list:
             # Delete the Glue table
             result = delete_glue_table(
-                data_product_name=self.data_product_config.name,
-                table_name=schema,
+                data_product_name=data_product_name,
+                table_name=schema_name,
                 logger=self.logger,
             )
             self.logger.info(str(result))
 
             # Delete a given elements raw and curated data for all versions of the data product
             delete_all_element_version_data_files(
-                data_product_name=self.data_product_config.name, table_name=schema
+                data_product_name=data_product_name, table_name=schema_name
             )
 
         current_metadata["schemas"] = [
             schema for schema in current_schemas if schema not in schema_list
         ]
         updated_metadata = DataProductMetadata(
-            data_product_name=self.data_product_config.name,
+            data_product_name=data_product_name,
             logger=self.logger,
             input_data=current_metadata,
         ).load()
@@ -141,27 +147,22 @@ class VersionCreator:
             self.logger.error(error)
             raise InvalidUpdate(error)
 
-        latest_version = updated_metadata.version
         new_version = generate_next_version_string(
-            latest_version, UpdateType.MajorUpdate
+            self.latest_version, UpdateType.MajorUpdate
         )
         self.logger.info(f"new version: {new_version}")
 
-        source_folder = f"{self.data_product_config.name}/{latest_version}/"
         # Copy files to the new version
-        s3_copy_folder_to_new_folder(
-            updated_metadata.write_bucket,
-            source_folder,
-            latest_version,
-            new_version,
-            self.logger,
+        self._copy_from_previous_version(
+            latest_version=self.latest_version, new_version=new_version
         )
+
         # Remove schema files that we no longer require in this version
         for schema in schema_list:
             # Get the current version of the schema path
-            schema_path = DataProductConfig(
-                name=self.data_product_config.name
-            ).schema_path(table_name=schema, version=new_version)
+            schema_path = DataProductConfig(name=data_product_name).schema_path(
+                table_name=schema, version=new_version
+            )
             # Delete the schema.json file for the table we have removed
             s3_client.delete_object(Bucket=schema_path.bucket, Key=schema_path.key)
 
@@ -172,17 +173,114 @@ class VersionCreator:
 
         return new_version
 
-    def update_metadata(self, input_data) -> str:
+    def _minor_version_bump(self, metadata: DataProductMetadata) -> str:
+        """Increment a Data Product by a major version, doing the following:
+            - increment version number
+            - copy all metadata from the previous version (metadata.json, schema.json)
+            - overwrite copied metadata.json with new version of metadata
+
+        Args:
+            metadata (DataProductMetadata): validated metadata
+
+        Returns:
+            str: new version number
         """
-        Create a new version with updated metadata.
+        new_version = generate_next_version_string(
+            self.latest_version, UpdateType.MinorUpdate
+        )
+        self._copy_from_previous_version(
+            latest_version=self.latest_version, new_version=new_version
+        )
+        new_version_key = self.data_product_config.metadata_path(new_version).key
+        metadata.write_json_to_s3(new_version_key)
+
+        return new_version
+
+    def _verify_input_metadata(self, input_data: dict) -> DataProductMetadata:
+        """Load and verify input Data Product metadata, raising errors if metadata is
+        missing, invalid, or if there are changes to fields that cannot be altered
+        between versions.
+
+        Args:
+            input_data (dict): input metadata in a format ready for loading into a
+            DataProductMetadata object.
+
+        Raises:
+            InvalidUpdate: raised if metadata is missing, invalid, or if there are
+            changes to fields that cannot be altered between versions
+
+        Returns:
+            DataProductMetadata: verified metadata object with current metadata in
+            metadata.data and previous metadata in metadata.latest_version_saved_data
         """
         metadata = DataProductMetadata(
             data_product_name=self.data_product_config.name,
             logger=self.logger,
             input_data=input_data,
         ).load()
+
         if not metadata.valid or not metadata.exists:
-            raise InvalidUpdate()
+            msg = "invalid metadata passed"
+            raise InvalidUpdate(msg)
+
+        changed_fields = metadata.changed_fields()
+        if changed_fields.difference(UPDATABLE_METADATA_FIELDS):
+            changed_fields = list(changed_fields)
+            num_fields = len(changed_fields)
+            msg = f"Non-updatable metadata field{('s'[:num_fields!=1])} changed:"
+            for f in changed_fields:
+                msg += f"{f}: {metadata.latest_version_saved_data[f]} -> {metadata.data[f]}; "
+            self.logger.error(msg)
+            raise InvalidUpdate(msg)
+
+        return metadata
+
+    def _verify_input_schema(
+        self, input_data: dict, table_name: str
+    ) -> DataProductSchema:
+        """Load and verify input Data Product schema, raising errors if schema is
+        missing, invalid, or if there is no associated parent Data Product.
+
+        Args:
+            input_data (dict): input schema in a format ready for loading into a
+            DataProductSchema object.
+
+        Raises:
+            InvalidUpdate: raised if schema is missing, invalid, or if there is no
+            associated parent Data Product
+
+        Returns:
+            DataProductSchema: verified schema object
+        """
+        schema = DataProductSchema(
+            data_product_name=self.data_product_config.name,
+            table_name=table_name,
+            logger=self.logger,
+            input_data=input_data,
+        ).load()
+
+        if not schema.valid:
+            error_message = (
+                f"schema for {table_name} has failed validation with the following error: "
+                f"{schema.error_traceback}"
+            )
+            raise InvalidUpdate(error_message)
+
+        if not schema.has_registered_data_product:
+            error_message = (
+                f"Schema for {table_name} has no registered metadata for the data "
+                "product it belongs to. Please first register the data product "
+                "metadata using the POST method of the /data-product/register endpoint."
+            )
+            raise InvalidUpdate(error_message)
+
+        return schema
+
+    def update_metadata(self, input_data) -> str:
+        """
+        Create a new version with updated metadata.
+        """
+        metadata = self._verify_input_metadata(input_data)
 
         state = metadata_update_type(metadata)
         self.logger.info(f"Update type {state}")
@@ -190,39 +288,24 @@ class VersionCreator:
         if state != UpdateType.MinorUpdate:
             raise InvalidUpdate(state)
 
-        latest_version = metadata.version
-        new_version = generate_next_version_string(latest_version, state)
-        self._copy_from_previous_version(
-            latest_version=latest_version, new_version=new_version
-        )
-        new_version_key = self.data_product_config.metadata_path(new_version).key
-        metadata.write_json_to_s3(new_version_key)
+        new_version = self._minor_version_bump(metadata)
 
         return new_version
 
-    def update_schema(self, input_data: dict, table_name: str):
+    def update_schema(self, input_data: dict, table_name: str) -> tuple[str, dict]:
         """
         Create a new version with updated schema.
         """
-        schema = DataProductSchema(
-            data_product_name=self.data_product_config.name,
-            table_name=table_name,
-            input_data=input_data,
-            logger=self.logger,
-        ).load()
-
-        if not schema.valid:
-            raise InvalidUpdate()
+        schema = self._verify_input_schema(input_data=input_data, table_name=table_name)
 
         state, changes = schema_update_type(schema)
         self.logger.info(f"Update type {state}")
         if not state == UpdateType.Unchanged:
             self.logger.info(f"Changes to schema: {changes}")
 
-            latest_version = schema.version
             new_version = generate_next_version_string(schema.version, state)
             self._copy_from_previous_version(
-                latest_version=latest_version, new_version=new_version
+                latest_version=self.latest_version, new_version=new_version
             )
             new_version_key = (
                 DataProductConfig(schema.data_product_name)
@@ -234,6 +317,53 @@ class VersionCreator:
             return new_version, changes
         else:
             raise InvalidUpdate()
+
+    def create_schema(
+        self, input_data: dict, table_name: str
+    ) -> tuple[str, DataProductSchema]:
+        """
+        Generate a version number for a new schema. Returns "v1.0" if this is the first
+        schema associated with that Data Product, otherwise returns a version number
+        after a minor version bump.
+        """
+        schema = self._verify_input_schema(input_data=input_data, table_name=table_name)
+
+        if schema.exists:
+            error_message = (
+                f"v1 of this schema for table {table_name} already exists. You can upversion this schema if "
+                "there are changes from v1 using the PUT method of this endpoint. Or if this is a different "
+                "table then please choose a different name for it."
+            )
+            raise InvalidUpdate(error_message)
+
+        data_product_name = self.data_product_config.name
+        metadata_dict = schema.parent_data_product_metadata
+
+        metadata = self._verify_input_metadata(metadata_dict)
+        schema.convert_schema_to_glue_table_input_csv()
+
+        if (
+            self.latest_version == "v1.0"
+            and not schema.parent_product_has_registered_schema
+        ):
+            self.logger.info(
+                f"No existing schemas are associated with {data_product_name}; "
+                f"setting version to 'v1.0' for {schema.table_name}"
+            )
+
+            metadata.write_json_to_s3(metadata.latest_version_key)
+
+            new_version = "v1.0"
+        else:
+            new_version = self._minor_version_bump(metadata)
+
+        schema.write_json_to_s3(
+            DataProductConfig(data_product_name)
+            .schema_path(schema.table_name, new_version)
+            .key
+        )
+
+        return new_version, schema
 
     def _copy_from_previous_version(self, latest_version, new_version):
         bucket, source_folder = self.data_product_config.metadata_path(
@@ -258,10 +388,14 @@ def metadata_update_type(data_product_metadata: DataProductMetadata) -> UpdateTy
 
     changed_fields = data_product_metadata.changed_fields()
     if not changed_fields:
+        # NOTE: this does not capture all types of updates (e.g. changes to a column
+        # data type inside an existing schema)
         return UpdateType.Unchanged
     if changed_fields.difference(UPDATABLE_METADATA_FIELDS):
         return UpdateType.NotAllowed
     else:
+        # NOTE: this does not capture all types of updates (e.g. removal of schema ->
+        # MajorUpdate)
         return UpdateType.MinorUpdate
 
 
@@ -305,8 +439,10 @@ def schema_update_type(
         }
         if not changed_fields:
             update_type = UpdateType.Unchanged
-        elif changed_fields.intersection(MINOR_UPDATE_SCHEMA_FIELDS):
+        elif changed_fields.intersection(MINOR_UPDATE_SCHEMA_FIELDS) == changed_fields:
             update_type = UpdateType.MinorUpdate
+        else:
+            update_type = UpdateType.MajorUpdate
 
     # could be returned and used to form notification of change to consumers of data when the
     # notification process is developed
@@ -337,7 +473,7 @@ def s3_copy_folder_to_new_folder(
     bucket, source_folder, latest_version, new_version, logger
 ):
     """
-    Recurisvely copy a folder, replacing {latest_version} with {new_version}
+    Recursively copy a folder, replacing {latest_version} with {new_version}
     """
     s3_client = boto3.client("s3")
     paginator = s3_client.get_paginator("list_objects_v2")