+++ terraform/aws/analytical-platform-production/cluster/terraform.tf
@@ -14,7 +14,7 @@ terraform {
     }
     aws = {
       source  = "hashicorp/aws"
-      version = "5.25.0"
+      version = "5.26.0"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"